execute_shift.v: restructure shifter
The original shifter code selects shifter results using nested ternary statements. This leads to a long chain of conditionals. This is now restructured to using case statements, which are "wider", but less "deep".

For now, both the origina code and the restructured code are included and can be selected via defines. In the future, the unused code path can be removed.