sound: Updated the DSP write port status busy flag (#118)
Previously the write port status busy flag was set on any DMA
request, and then cleared when the software polled for the
status. This caused problems with software using auto-init DMA
because the busy flag may be reset by the new transfer before
it polled a second time resulting in the DSP always returning
busy.

The logic was updated to instead use a fixed busy timer that
clears independently from the software polling for status.